requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.7.4
cffi==1.17.1
channels==4.1.0
channels_redis==4.2.1
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
decorator==5.1.1
Django==5.1
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
msgpack==1.1.0
mysql-connector-python==9.0.0
openrouteservice==2.3.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.3.0
python-dotenv==1.0.1
ratelim==0.1.6
redis==5.2.0
requests==2.32.3
service-identity==24.2.0
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.2
websockets==13.1
zope.interface==7.2
